requirements.txt
python-dotenv
flask
twilio
openai